refactor: start with new `SecClient` definition
