+++ sec/src/lib/implementations/states/extract/execute_sec_request/data/input/mod.rs
@@ -28,8 +28,8 @@
 use std::fmt;
 
 use crate::error::State as StateError;
-use crate::shared::sec_client::SecClient;
-use crate::shared::sec_client::traits::sec_client::SecClient as SecClientTrait;
+use crate::shared::old_sec_client::SecClient;
+use crate::shared::old_sec_client::traits::sec_client::SecClient as SecClientTrait;
 use crate::shared::sec_request::SecRequest;
 use crate::shared::sec_request::implementations::reqwest_request::ReqwestRequest;
 use crate::shared::sec_request::traits::inner_request::InnerRequest;
