+++ sec/src/lib/implementations/states/extract/mod.rs
@@ -49,7 +49,7 @@ use crate::implementations::states::extract::validate_cik_format::{
 use crate::implementations::states::extract::validate_sec_response::ValidateSecResponse;
 
 use crate::shared::cik::Cik;
-use crate::shared::sec_client::SecClient;
+use crate::shared::old_sec_client::SecClient;
 use crate::shared::sec_request::SecRequest;
 use crate::shared::sec_request::implementations::reqwest_request::ReqwestRequest;
