+++ sec/src/lib/implementations/states/extract/prepare_sec_request/data/output/mod.rs
@@ -31,7 +31,7 @@ use std::{fmt, hash::Hash};
 use state_maschine::prelude::StateData as SMStateData;
 
 use crate::error::State as StateError;
-use crate::shared::sec_client::SecClient;
+use crate::shared::old_sec_client::SecClient;
 use crate::shared::sec_request::SecRequest;
 use crate::shared::sec_request::implementations::reqwest_request::ReqwestRequest;
 use crate::shared::sec_request::traits::inner_request::InnerRequest;
@@ -215,7 +215,7 @@ mod tests {
 
     use super::{PrepareSecRequestOutput, PrepareSecRequestOutputUpdaterBuilder};
     use crate::shared::cik::Cik;
-    use crate::shared::sec_client::SecClient;
+    use crate::shared::old_sec_client::SecClient;
     use crate::shared::sec_request::SecRequest;
     use crate::shared::user_agent::UserAgent;
     use crate::traits::state_machine::state::StateData;
