+++ sec/src/lib/shared/mod.rs
@@ -22,6 +22,7 @@
 pub mod cik;
 pub mod http_client;
 pub mod json_response;
+pub mod old_sec_client; // TODO: DELETE AFTER new sec_client definition is implemented
 pub mod sec_client;
 pub mod sec_request;
 pub mod sec_response;
